Modules.IoxideFiles: one class per file
Split the single file into IoxideFiles / IoxideFilesBuilder / IoxideFilesHandler / IoxideAssetContent,
matching GenHTTP's module conventions.
